Provide a way to clean up DSDict keys when dsdict_destroy is called by providing a key free function.
Previously keys would not be freed (despite the fact that DSDict's buckets hold a reference to the key), but this meant that callers would be required to hold onto a reference to all of their keys. Now, the key reference will be destroyed at the same time as the DSDict if the caller specifies a key free function at instantiation.